+++ dashboard-api/internal/study_segments.go
@@ -17,13 +17,13 @@ type SegmentProgress struct {
 	Id                          string  `json:"id"`
 	Name                        string  `json:"name"`
 	Datetime                    int64   `json:"datetime"`
-	CurrentBudget               int64   `json:"currentBudget"`
+	CurrentBudget               float64 `json:"currentBudget"`
 	DesiredPercentage           float64 `json:"desiredPercentage"`
 	CurrentPercentage           float64 `json:"currentPercentage"`
 	ExpectedPercentage          float64 `json:"expectedPercentage"`
 	DesiredParticipants         *int64  `json:"desiredParticipants"`
 	ExpectedParticipants        int64   `json:"expectedParticipants"`
 	CurrentParticipants         int64   `json:"currentParticipants"`
-	CurrentPricePerParticipant  int64   `json:"currentPricePerParticipant"`
+	CurrentPricePerParticipant  float64 `json:"currentPricePerParticipant"`
 	PercentageDeviationFromGoal float64 `json:"percentageDeviationFromGoal"`
 }
